+++ src/main/java/za/co/hpsc/web/constants/IpscConstants.java
@@ -38,5 +38,6 @@ public final class IpscConstants {
     public static final int PERCENTAGE_SCALE = 2;
 
     public static final MatchCategory DEFAULT_MATCH_CATEGORY = MatchCategory.CLUB_SHOOT;
-    public static final ClubIdentifier DEFAULT_CLUB_IDENTIFIER = ClubIdentifier.HPSC;
+    public static final ClubIdentifier DEFAULT_MATCH_CLUB_IDENTIFIER = ClubIdentifier.ALL;
+    public static final ClubIdentifier HOME_CLUB_IDENTIFIER = ClubIdentifier.HPSC;
 }